fix: variables types

+++ proj2/stat_display.cpp
@@ -1,6 +1,8 @@
 #include "stat_display.h"
 
-void print_stat(long int duration, long long unsigned int max_seq, double data_bits, int data_pkts, bool isRcv, double avg_delay, double min_delay, double max_delay, int re_pkts)
+void print_stat(double duration, long long unsigned int max_seq, long long unsigned int data_bits, 
+    long long unsigned int data_pkts, bool isRcv, double avg_delay, double min_delay, double max_delay, 
+        long long unsigned int re_pkts)
 {
     double rate = data_bits / (1000000 * duration);
     double rate_pkt = data_pkts / (duration);
